requirements.txt
# =============================================
#  Pinned versions matching the tested environment.
#  Install with:  pip install -r requirements.txt
# =============================================

# --- Core ML ---
torch==2.8.0
torchaudio==2.8.0
pytorch-lightning==2.2.5

# --- Configuration ---
hydra-core==1.3.2
omegaconf==2.3.0
hydra-colorlog==1.2.0
antlr4-python3-runtime==4.9.3

# --- Audio processing ---
librosa==0.10.2.post1
soundfile==0.12.1
scipy==1.13.1
pystoi==0.4.1
pesq==0.0.4

# --- Data ---
h5py==3.11.0
numpy==1.26.4

# --- Visualization ---
matplotlib==3.9.0

# --- Room acoustics (only needed for data generation) ---
pyroomacoustics==0.7.4
# rir-generator==0.2.0  # uncomment if generating RIR data

# --- Misc ---
tqdm==4.67.1
colorlog==6.9.0
